+++ packages/expo-modules-core/ios/Core/Views/SwiftUI/SwiftUIViewDefinition.swift
@@ -32,21 +32,41 @@ extension ExpoSwiftUI {
    A definition representing the native SwiftUI view to export to React.
    */
   public final class ViewDefinition<Props: ViewProps, ViewType: View<Props>>: ExpoModulesCore.ViewDefinition<HostingView<Props, ViewType>> {
+    // To obtain prop and event names from the props object we need to create a dummy instance first.
+    // This is not ideal, but RN requires us to provide all names before the view is created
+    // and there doesn't seem to be a better way to do this right now.
+    private lazy var dummyPropsMirror: Mirror = Mirror(reflecting: Props())
+
     init(_ viewType: ViewType.Type) {
       super.init(HostingView<Props, ViewType>.self, elements: [])
     }
 
     public override func createView(appContext: AppContext) -> UIView? {
       let props = Props()
-      return HostingView(viewType: ViewType.self, props: props, appContext: appContext)
+      let view = HostingView(viewType: ViewType.self, props: props, appContext: appContext)
+
+      // Set up events to call view's `dispatchEvent` method.
+      props.setUpEvents(view.dispatchEvent(_:payload:))
+
+      return view
     }
 
     public override func getSupportedPropNames() -> [String] {
-      // To obtain field names from the props object we need to create a dummy instance first.
-      // This is not ideal, but RN requires us to provide all prop names before the view is created
-      // and there doesn't seem to be a better way to do this right now.
-      let props = Props()
-      return fieldsOf(props).compactMap(\.key)
+      return dummyPropsMirror.children.compactMap { (label: String?, value: Any) in
+        guard let field = value as? AnyFieldInternal else {
+          return nil
+        }
+        return field.key ?? convertLabelToKey(label)
+      }
+    }
+
+    public override func getSupportedEventNames() -> [String] {
+      return dummyPropsMirror.children.compactMap { (label: String?, value: Any) in
+        guard let event = value as? EventDispatcher else {
+          return nil
+        }
+        return event.customName ?? convertLabelToKey(label)
+      }
     }
   }
 }